Massive refactor of sample manifest reception.
The issue originally started that all of the samples in a manifest were
then requested in parallel, which for a 3000 entry manifest caused pain.
This was an attempt to restructure the code in such a way to make that
far more friendly, but degenerated into fixing no end of bugs & issues.

Collect errors per tube.
Collect errors globally for missing tubes.
Global functions to deal with the various stages of processing.
Permit manifest upload even if there are errors, with a warning.
Hide controls in the face of success.