Bind studies to manifest templates.
Manifest templates are really about supplying studies.  Hence we bind
them to a specific group of studies.  We currently have one template
feeding multiple studies, but it is possible to have multiple templates
feeding a study.

Added a couple of dummy studies into the plate manifest purely to
demonstrate the functionality.